requirements.txt
blinker==1.4
boto3==1.17.33
botocore==1.20.33
cachelib==0.1.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Session==0.3.2
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
inflect==5.3.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-dotenv==0.15.0
redis==3.5.3
requests==2.25.1
s3transfer==0.3.6
sentry-sdk==1.0.0
six==1.15.0
soupsieve==2.2.1
stemming==1.0.1
urllib3==1.26.4
Werkzeug==1.0.1
wikipedia==1.4.0
WTForms==2.3.3
pandas==1.4.2
